Hacked version of contains? from /u/flatmapcatmap
In the Reddit discussion on the Clojure forum,
this change to a fast-contains improves Clojure's speed by almost
2x. Quoting /u/flatmapcatmap:

"...clojure.lang.util.Equiv seems to be incredibly slow. If I replace the
above call to .equals with =, the run takes about 3x as long!"

Hoping some Clojure guru will chime-in and explain why Clojure
fails so hard (speed-wise) when using 'contains?'.